Fully supports int and char

+++ inc/libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: aouahib <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/23 17:52:22 by aouahib           #+#    #+#             */
-/*   Updated: 2019/10/30 11:54:22 by aouahib          ###   ########.fr       */
+/*   Updated: 2019/10/30 15:47:16 by aouahib          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,9 +35,10 @@ typedef struct	s_printf
 int				ft_printf(const char *format, ...);
 int				pf_isflag(char c);
 int				pf_istype(char c);
-int				pf_getdignum(int n, int base);
+int				pf_getdignum(long long n, int base);
 t_printf		*pf_parse(const char **format, va_list *vl);
 int				pf_putstr(t_printf *pf, va_list *vl);
 int				pf_putint(t_printf *pf, va_list *vl);
 int				pf_puthex(t_printf *pf, va_list *vl);
+
 #endif
